Go back to right commit in pandas#

+++ pandas/pandas/tests/frame/methods/test_asfreq.py
@@ -159,8 +159,7 @@ class TestAsFreq:
 
         # setup
         rng = date_range("1/1/2016", periods=10, freq="2S")
-        # Explicit cast to 'float' to avoid implicit cast when setting None
-        ts = Series(np.arange(len(rng)), index=rng, dtype="float")
+        ts = Series(np.arange(len(rng)), index=rng)
         df = DataFrame({"one": ts})
 
         # insert pre-existing missing value
